refactor(runtime): extract bounded gate module
